requirements.txt
discord.py
python-dotenv
requests
praw
asyncio